refactor: branch depending on context, then on permission
