+++ src/components/App.js
@@ -1,29 +1,28 @@
 import React, { useContext } from 'react';
-import TeacherMode from './modes/teacher/TeacherMode';
-import StudentMode from './modes/student/StudentMode';
 import ModalProviders from './context/ModalProviders';
-import { DEFAULT_PERMISSION, PERMISSION_LEVELS } from '../config/constants';
 import { Context } from './context/ContextContext';
 import { TokenProvider } from './context/TokenContext';
+import CONTEXTS from '../config/contexts';
+import BuilderView from './views/BuilderView';
+import PlayerView from './views/PlayerView';
 
 const App = () => {
   const context = useContext(Context);
 
   const renderContent = () => {
-    switch (context?.get('permission', DEFAULT_PERMISSION)) {
-      case PERMISSION_LEVELS.WRITE:
-      case PERMISSION_LEVELS.ADMIN:
+    switch (context?.get('context')) {
+      case CONTEXTS.BUILDER:
         return (
           <ModalProviders>
-            <TeacherMode />
+            <BuilderView />
           </ModalProviders>
         );
 
-      case PERMISSION_LEVELS.READ:
+      case CONTEXTS.PLAYER:
       default:
         return (
           <ModalProviders>
-            <StudentMode />
+            <PlayerView />
           </ModalProviders>
         );
     }
